requirements.txt
absl-py==2.3.1
accelerate==1.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
albucore==0.0.24
albumentations==2.0.8
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
beartype==0.22.2
bitarray==3.0.0
blinker==1.9.0
braceexpand==0.1.7
certifi==2025.10.5
cffi==1.17.1
charset-normalizer==3.4.3
click==8.1.7
coloredlogs==15.0.1
comm==0.2.3
contourpy==1.3.1
cosine_annealing_warmup @ git+https://github.com/katsura-jp/pytorch-cosine-annealing-with-warmup@12d03c07553aedd3d9e9155e2b3e31ce8c64081a
cycler==0.12.1
Cython==3.0.11
datasets==3.2.0
debugpy==1.8.17
decorator==5.2.1
diffusers==0.32.1
dill==0.3.8
docker-pycreds==0.4.0
easydict==1.13
efficientnet_pytorch==0.7.1
einops==0.8.0
ema-pytorch==0.7.7
et_xmlfile==2.0.0
exceptiongroup==1.3.0
executing==2.2.1
fairseq_signals_backbone @ git+https://github.com/fuying-wang/fairseq-signals@27d94bab8a1040879c011609df1488aac21a586a
filelock==3.20.0
flatbuffers==25.9.23
fonttools==4.55.1
frozenlist==1.5.0
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
grpcio==1.75.1
h5py==3.14.0
hf-xet==1.1.10
huggingface-hub==0.34.4
humanfriendly==10.0
hydra-core==1.3.2
idna==3.10
imageio==2.37.0
importlib_metadata==8.7.0
insightface==0.7.3
ipdb==0.13.13
ipykernel==6.30.1
ipython==8.37.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.7
kornia==0.8.1
kornia_rs==0.1.9
lazy_loader==0.4
lightning==2.4.0
lightning-utilities==0.11.9
llvmlite==0.46.0
loguru==0.7.3
lxml==5.3.0
Markdown==3.9
MarkupSafe==3.0.3
matplotlib==3.9.3
matplotlib-inline==0.1.7
ml_dtypes==0.5.3
mne==1.10.1
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
munch==4.0.0
narwhals==2.6.0
nest-asyncio==1.6.0
networkx==3.4.2
neurokit2==0.2.12
ninja==1.13.0
nltk==3.9.1
numba==0.63.1
numpy==2.1.2
omegaconf==2.3.0
onnx==1.19.1
onnx2torch==1.5.15
onnxruntime==1.23.1
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
parso==0.8.5
peft==0.14.0
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.4.0
pooch==1.8.2
portalocker==3.0.0
POT==0.9.5
pretrainedmodels==0.7.4
prettytable==3.16.0
prompt_toolkit==3.0.52
propcache==0.2.1
protobuf==5.29.1
psutil==7.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==18.1.0
pycparser==2.23
pydantic==2.10.3
pydantic_core==2.27.1
pydeck==0.9.1
Pygments==2.19.2
pynndescent==0.5.13
pyparsing==3.2.0
pysam==0.23.3
python-dateutil==2.9.0.post0
pytorch-lightning==2.4.0
pytorch-warmup==0.2.0
pytz==2024.2
PyWavelets==1.8.0
PyYAML==6.0.3
pyzmq==27.1.0
regex==2024.11.6
requests==2.32.5
sacrebleu==2.4.3
safetensors==0.6.2
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.14.1
seaborn==0.13.2
segmentation_models_pytorch==0.4.0
sentencepiece==0.2.1
sentry-sdk==2.19.2
setproctitle==1.3.4
simsimd==6.5.3
six==1.17.0
smmap==5.0.1
soundfile==0.12.1
stack-data==0.6.3
streamlit==1.50.0
stringzilla==4.2.1
sympy==1.13.1
tabulate==0.9.0
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.4
threadpoolctl==3.5.0
tifffile==2025.5.10
timm==1.0.12
tokenizers==0.21.0
toml==0.10.2
tomli==2.2.1
torch==2.5.1
torchaudio==2.5.1
torchdiffeq==0.2.5
torchmetrics==1.6.0
torchtools @ git+https://github.com/pabloppp/pytorch-tools@610158d5016d6418aee27f956e7afd17ff35ba04
torchvision==0.20.1
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.47.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2024.2
umap-learn==0.5.9.post2
unet==0.8.1
urllib3==2.5.0
vitaldb==1.5.8
wandb==0.21.4
warmup_scheduler==0.3
watchdog==6.0.0
wcwidth==0.2.13
webdataset==1.0.2
Werkzeug==3.1.3
wfdb==4.1.2
xgboost==2.1.3
xxhash==3.5.0
yarl==1.18.3
zipp==3.23.0
